==> script/Analysis 2024 for Forestry/6-Plot.Rmd <==

```{r include=FALSE}
library(tidyverse)
library(readxl)
library(writexl)
library(sf)
library(here)

library(flextable)
library(ftExtra)
library(showtext)

showtext_auto()
font_add("Microsoft JhengHei", "msjh.ttc")
here::here() 
```

```{r include=FALSE}
order_Office <- c("宜蘭","新竹","臺中","南投", "嘉義","屏東", "花蓮","臺東", "Total")
```



```{r include=FALSE}

M.data <- 
  here("./data/clean/Forestry/for analysis/") %>% 
  list.files(., full.names = T) %>% 
  str_subset(paste0(2020:2024)) %>% #如果只要2020~2022年，就寫2020:2022
  lapply(., read_excel, sheet="Data", col_types = "text") %>% 
  bind_rows() %>% 
  
  mutate(Office = str_replace_all(Office, c("羅東" = "宜蘭",
                                            "東勢" = "臺中"))) %>%


  # mutate(Office = ordered(Office, 
  #                         c("羅東", "新竹", "東勢", "南投",
  #                           "嘉義", "屏東", "花蓮", "臺東"))) %>% 
  # mutate(TypeName.1 = ordered(TypeName.1,
  #                             c("闊葉林", "針葉林", "竹林",
  #                               "混淆林", "非森林"))) %>% 
  mutate_at(c("Year", "Survey","Month",
              "Day", "Macaca_sur", "Distance", "Altitude", "julian.D"), as.numeric) 
  
```

### Part 1 最初的資料

    說明：收到的資料，尚未刪減任何資料。  
     - 包括有回報因故無法調查的樣區。

各分署的調查人員人數

```{r 調查者的統計資料, echo=FALSE, message=FALSE, warning=FALSE}

#調查者的統計資料
List_surveyor <- 
  M.data %>%
  filter(!is.na(Macaca_sur)) %>% 
  separate(.,Surveyor,
           into = paste0("Surveyor","_",0:10),
           sep ="、|,",
           extra = "drop",
           fill = "right") %>% 
  reshape2::melt(., 
                 id.vars = c("Office","Year", "Site_N",
                             "Point",  "Survey"),
                 measure.vars = paste0("Surveyor","_",0:10),
                 variable.name = "Surveyor",
                 value.name = "Name")%>% 
  filter(!is.na(Name)) 

List_surveyor %>% 
  mutate(SP = paste0(Site_N, "-", Point)) %>% 
  group_by(Year, Office) %>% 
  summarise(Site_n = Site_N %>% unique %>% length,     #樣區數
            Person_n = Name %>% unique %>% length) %>%  #人數
  arrange(Office)%>% 
  ungroup() %>% 
 # select(Office, Person_n) %>% 
  bind_rows(List_surveyor %>%
              group_by(Year) %>% 
              summarise(Person_n = Name %>%
                          unique %>% length,
                        Site_n = Site_N %>%
                          unique %>% length,) %>%
              
              mutate(Office = "Total") ) %>% 
  mutate(Year = as.character(Year)) %>% 
  
  filter(Year %in% c(2023,2024)) %>% 
  reshape2::melt(id = 1:2 )%>% 
  reshape2::dcast(Office ~ Year + variable, var.value = "value") %>% 
  mutate(Office = ordered(Office, order_Office)) %>% 
  arrange(Office) %>% 
  flextable(.) %>% 
   width(., width = rep(1,5)) %>% 


  set_header_labels(
    'Office' = "分署",
    '2023_Site_n' = "樣區數",
    '2023_Person_n' = "調查者人數",
    '2024_Site_n' = "樣區數",
    '2024_Person_n' = "調查者人數" ) %>% 
  add_header_row(., 
                 values = c(" ", "2023", "2024"),
                 colwidths = c(1,2,2)) %>% 
  
  align(.,j = 1:5, align = "center", part = "all") 

```


```{r}
M.data %>% 
  filter(! is.na(Macaca_sur)) %>% 
  group_by(Office,Site_N, Year) %>% 
  summarise(
    # Surveyor_n = str_split(Surveyor, "、") %>%
    #   c(.) %>% unique(.) %>% 
    #   paste(., collapse = "、"),
    Point_n = n(),
    Point = Point %>% unique() %>% length() %>%length
  ) %>% View
  group_by(Office, Year) %>% 
  summarise(
    Point_n = sum(Point_n),
    Point = sum(Point)
  ) %>% 
bind_rows(
  group_by(.,Year) %>% 
    summarise_at(c("Point_n", "Point"), sum) %>% 
    mutate(Office = "Total")
) %>% 
  filter(Year %in% 2024)
```



<br/>


今年沒調查的樣區列表
```{r echo=FALSE}

M.data %>% 
  mutate(Year = ifelse(is.na(Year), "2020", Year)) %>% 
  filter(Year %in% c("2024")) %>% 
  filter(is.na(Macaca_sur)) %>% 
  reshape2::dcast(Site_N + Name ~Year ,
                  length, value.var = "Macaca_sur") %>% 
  arrange(str_sub(Site_N, 4,6)) %>% 
  flextable() %>% 
  autofit()


```


\newpage

### Part 2 刪疏失的資料

    刪除項目：
    1. 同一旅次超過7日才完成調查：同一旅次同一樣區內超過7日才完成調查，整個旅次的資料放棄。
    2. 開始時間晚於10點54分。
    3. 調查未滿6分鐘刪。
    4. 不在預設樣點上：容許GPS誤差在50 m，調查位置座標離預設樣點距離>50m刪除。

整體的樣點次流變

```{r echo=FALSE, message=FALSE, warning=FALSE}

M.data %>% 
  filter(! is.na(Macaca_sur)) %>% 
  group_by(Year) %>% 
  summarise(
    '1 收到' = n(),
    '2 檢核後' = str_subset(analysis, "locate|6min|Toolate|7day|month", negate = T) %>% length,
    '2-1 不足6分鐘' = str_subset(analysis, "6min") %>% length,
    '2-2 晚於11時才完成調查' = str_subset(analysis, "Toolate") %>% length,
    '2-3 超過7日' = str_subset(analysis, "7day") %>% length,
    '2-4 位置錯誤' = str_subset(analysis, "locate") %>% length,
    '2-5 3~6月以外' = str_subset(analysis, "month") %>% length,
    
    '3 刪掉低於50m(3~6月、<50m、所有棲地類型)' =
      str_subset(analysis, "^Y")%>% 
      str_subset(., "month", negate = T) %>% 
      str_subset(., "50m", negate = T) %>% length,

    '4 篩選後(3~6月、=>50m、only森林)' = str_subset(analysis, "^Y$") %>% length
  ) %>% 
  filter(Year %in% c(2023,2024)) %>% 
  
  reshape2::melt(id = 1 )%>% 
  reshape2::dcast(variable ~ Year , var.value = "value")%>% 
  flextable(.) %>% 
   width(., width = c(3,1,1)) %>% 
  
#  align(.,j = 2,
#        align = "center", part = "all") %>% 
   padding(., i=3:7, j=1, padding.left = 20)


```

<br/>

\newpage

樣區樣點的統計資料

```{r echo=FALSE, message=FALSE, warning=FALSE}
M.data %>% 
  filter(! is.na(Macaca_sur)) %>% 
  filter(Year == 2023) %>% 
  group_by(Office, Survey) %>% 
  summarise(
    Point_n = n(),
    Good_n = str_subset(analysis, "^Y$|\\d{2}m|\\({1}\\w{9}\\)") %>%length
  ) %>% 
  ungroup() %>%
  

  
  
  bind_rows(
    M.data %>% 
   filter(Year == 2024) %>% 
      filter(! is.na(Macaca_sur))%>% 
      group_by(Survey) %>% 
      summarise(
        Point_n = n(),
        Good_n = str_subset(analysis, "^Y$|\\d{2}m|\\({1}\\w{9}\\)") %>%length
        ) %>% 
      mutate(Office = "Total") %>% 
      ungroup() 
    ) %>% 
  
  mutate(prob = (100 * Good_n/Point_n) %>%  round(1)) %>%
  reshape2::melt(id = 1:2) %>% 
  reshape2::dcast(Office ~ Survey+variable) %>% 
  mutate(Office = ordered(Office, order_Office)) %>% 
  arrange(Office) %>% 
  flextable(.) %>% 

  set_header_labels(
    "Office"="分署",
    "1_Point_n"="樣點",
    "1_Good_n" = "無疏失的樣點",
    "1_prob" = "正確率%",
    "2_Point_n"="樣點",
    "2_Good_n" = "無疏失的樣點",
    "2_prob" = "正確率%") %>% 
  add_header_row(., 
                 values = c(" ","第1旅次", "第2旅次"),
                 colwidths = c(1,3,3)) %>% 

  theme_booktabs(.)%>% 
  align(.,
        align = "center", part = "header") 
  
```



<br/>\
<br/>

```{r echo=FALSE, message=FALSE, warning=FALSE}
M.data %>% 
  filter(! is.na(Macaca_sur)) %>% 
  group_by(Office, Year) %>% 
  summarise(
    Point_n = n(),
    Good_n = str_subset(analysis, "^Y$|\\d{2}m|\\({1}\\w{9}\\)") %>%length
  ) %>% 
  ungroup() %>%
  

  
  bind_rows(
    M.data %>% 

      filter(! is.na(Macaca_sur))%>% 
      group_by(Year) %>% 
      summarise(
        Point_n = n(),
        Good_n = str_subset(analysis, "^Y$|\\d{2}m|\\({1}\\w{9}\\)") %>%length
        ) %>% 
      mutate(Office = "Total") %>% 
      ungroup() 
    ) %>% 
  
  mutate(prob = (100 * Good_n/Point_n) %>%  round(1)) %>%
  reshape2::melt(id = 1:2) %>% 
  reshape2::dcast(Office ~ Year + variable, value.var = "value") %>%
  mutate(level2 = ifelse(Office %in% "Total", 2, 1)) %>% 

  arrange(level2, desc(`2024_prob`) )%>% 
  
  mutate(prob_diff = paste0(`2024_prob`, "(",
                            round( `2024_prob`-`2023_prob`, 1), ")")
         ) %>% 
  
  dplyr::select(-`level2`, -starts_with(c("2020","2021","2022","2023")),
         -`2024_prob`) %>% 
  flextable(.)  %>% 
  
  set_header_labels(
    "Office"="分署",
    "2024_Point_n"="樣點",
    "2024_Good_n" = "無疏失的樣點",
    "prob_diff" = "正確率%") %>% 
  add_header_row(., 
                 values = c(" ", "2024"),
                 colwidths = c(1,3)) %>% 
  

  theme_booktabs(.)%>% 
  align(.,
        align = "center", part = "all")
  


```

<br/>\
<br/>

\newpage

2024年各分署各疏失資料類型的筆數

```{r echo=FALSE, message=FALSE, warning=FALSE}

M.data %>% 
  filter(! is.na(Macaca_sur)) %>% 
  filter(Year == 2024) %>% 
  bind_rows(
    M.data %>% 
  filter(Year == 2024) %>%       
      filter(! is.na(Macaca_sur)) %>% 
      mutate(Office = "Total")
  ) %>% 
  mutate(Office = ordered(Office, order_Office)) %>% 
  group_by(Office, Survey) %>% 
  summarise(
    '樣點數' = n(),
    '超過7日' = str_subset(analysis, "7day") %>% length,
    '晚於11時' = str_subset(analysis, "Toolate") %>% length,
    '不足6分鐘' = str_subset(analysis, "6min") %>% length,
    '誤差超過50m' = str_subset(analysis, "locate") %>% length,
    '不在3~6月內' = str_subset(analysis, "month") %>% length
  ) %>% 
  mutate_all(~str_replace_all(., '^0$', "-")) %>% 
  
  flextable(.) %>% 
  align(.,
        align = "center", part = "all")
```

<br/>

\newpage

```{r echo=FALSE, message=FALSE, warning=FALSE}

M.data %>% 
  filter(! is.na(Macaca_sur)) %>% 
  bind_rows(
    M.data %>% 
      filter(! is.na(Macaca_sur)) %>% 
      mutate(Office = "Total")
  ) %>% 
  mutate(Office = ordered(Office, c("宜蘭","新竹",
                                    "臺中","南投",
                                    "嘉義","屏東",
                                    "花蓮","臺東",
                                    "Total"))) %>% 
  group_by(Office, Year) %>% 
  summarise(
    '超過7日' = str_subset(analysis, "7day") %>% length,
    '晚於11時' = str_subset(analysis, "Toolate") %>% length ,
    '不足6分鐘' = str_subset(analysis, "6min") %>% length ,
    '誤差超過50m' = str_subset(analysis, "locate") %>% length,
    '不在3~6月內' = str_subset(analysis, "month") %>% length
    
  ) %>% 


  mutate_all(~str_replace_all(., '^0$', "-")) %>% 
  reshape2::melt(id = 1:2) %>% 
  filter(Year %in% c(2023,2024)) %>% 
  reshape2::dcast(Office ~ Year + variable, value.var = "value") %>% 
  
  flextable(.) %>% 

  set_header_labels(
    "Office"="分署",

    "2023_超過7日" = "超過7日",
    "2023_晚於11時" = "晚於11時",
    "2023_不足6分鐘"="不足6分鐘",
    "2023_誤差超過50m" = "誤差超過50m",
    "2023_不在3~6月內" = "不在3~6月內",
    "2024_超過7日" = "超過7日",
    "2024_晚於11時" = "晚於11時",
    "2024_不足6分鐘"="不足6分鐘",
    "2024_誤差超過50m" = "誤差超過50m",
    "2024_不在3~6月內" = "不在3~6月內") %>%
  add_header_row(.,
                 values = c(" ", "2023", "2024"),
                 colwidths = c(1,5,5)) %>%
  align(.,
        align = "center", part = "all")
```

<br/>

\newpage

篩選後的資料後，各分署的獼猴調查統計

```{r echo=FALSE, message=FALSE, warning=FALSE}
M.data %>% 
  filter(Year == 2024) %>% 
  filter(analysis %in% "Y") %>% 
  bind_rows(
    M.data %>% 
      filter(! is.na(Macaca_sur)) %>% 
      filter(analysis %in% "Y") %>%
      filter(Year == 2024) %>% 
      mutate(Office = "Total")
  ) %>% 
  mutate(Office = ordered(Office, c("宜蘭","新竹",
                                    "臺中","南投",
                                    "嘉義","屏東",
                                    "花蓮","臺東",
                                    "Total"))) %>% 
  group_by(Office, Survey) %>% 
  summarise( 
    '樣點數' = n(), 
    '孤猴' = str_subset(Macaca_sur.ori, "1") %>% length, 
    '猴群' = str_subset(Macaca_sur.ori, "2") %>% length
  )%>% 
  
  flextable(.) %>% 
  align(.,
        align = "center", part = "all")
```

<br/>

\newpage

### Part 3 納入分析的資料

    不納入分析的情形：

    1. 移除非森林。樣點座標距離森林圖層>20m者為非森林。
    2. 移除海拔<50m的樣點(及猴群)資料。
    3. 移除3/1~6/30以前或以後的調查。
    4. >100m的猴群不納入分析。(猴群改為0)
    5. 同一旅次同一樣區內兩猴群所在樣點的距離<300者，僅留1群。

記錄到1\~5群猴群的樣區的樣區數

```{r echo=FALSE, message=FALSE, warning=FALSE}
M.data %>% 
  filter(Year == 2024) %>% 
  filter(analysis %in% "Y") %>% 
  filter(Macaca_sur.ori %in% 2 & !Macaca_dist %in% "C")%>% 
  group_by(Year, Survey, Site_N) %>% 
  summarise(N = n()) %>% 
  group_by(N) %>% 
  summarise(N_befoer = n()) %>% 
  
  left_join(
  M.data %>% 
    filter(Year == 2024) %>% 
  filter(analysis %in% "Y") %>% 
  filter(Macaca_sur %in% 1) %>% 
  group_by(Year, Survey, Site_N) %>% 
  summarise(N = n()) %>% 
   group_by(N) %>% 
  summarise(N_after = n()) 
    
  )%>% 
  mutate(N_after = ifelse(is.na(N_after), "-", N_after)) %>% 
  flextable(.)  %>% 
   width(., width = c(2,1.5,1.5)) %>% 
  align(.,j = 1:3,
        align = "center", part = "all")  %>%
  set_header_labels(
    'N' = "單一樣區內的猴群數",
    'N_befoer' = "樣區數 (刪除前)",
    'N_after' = "樣區數 (刪除後)" ) 
  
```

刪除重複記錄前猴群有`r  nrow(M.data[M.data$Year %in% 2024 & M.data$analysis %in% "Y" & M.data$Macaca_dist != "C" &M.data$Macaca_sur.ori %in% 2,])`群；刪除重複記錄後猴群有`r  sum(M.data[M.data$Year %in% 2024 & M.data$analysis %in% "Y",]$Macaca_sur)`群。\
<br/>\
<br/>\
<br/>

\newpage

2024年位在海拔50 m以上森林的調查樣點數、臺灣獼猴( *Macaca cyclopis* )的猴群數及相對密度

```{r echo=FALSE, message=FALSE, warning=FALSE}

E_rate_Office<-
    M.data %>%
  filter(Year == 2024) %>% 
    filter(analysis %in% "Y") %>% 
    group_by(Year, Survey, Office) %>%
    summarise(N = n(),
              m = sum(Macaca_sur)) %>% 
    bind_rows(group_by(.,Year, Survey) %>%  #增加total的部分
                summarise(N=sum(N), m = sum(m)) %>%
                mutate(Office='Total')) %>% 
    mutate(E = m/N) %>%
    ungroup() %>%
  
    group_by(Office) %>%
    summarise(Mean_N = mean(N),
              Se_N = sd(N)/sqrt(n()),
              Mean_m = mean(m),
              Se_m = sd(m)/sqrt(n()),
              Mean_E = mean(E),
              Se_E = sd(E)/sqrt(n())) 


E_rate_Office %>%
  mutate(Mean_E = round(Mean_E, 3),
         Se_E = round(Se_E, 3)) %>% 
  mutate(Office = ordered(Office, order_Office)) %>% 
  arrange(Office) %>% 
  flextable(.) %>%
  
  align(.,
        align = "center", part = "all")  
```

<br/>\
<br/>

\newpage

2024年各森林類型內的調查樣點數、臺灣獼猴( *Macaca cyclopis* )的猴群數及相對密度

```{r echo=FALSE, message=FALSE, warning=FALSE}
E_rate_Forest<-

#森林內的分別計
M.data %>%  
  filter(Year == 2024) %>% 
  filter(analysis %in% "Y") %>%
  group_by(Year, Survey, TypeName.1) %>% 
  summarise(N = n(),
            m = sum(Macaca_sur)) %>% 
  
  bind_rows(  #森林的總計
    M.data %>% 
      filter(Year == 2024) %>% 
      filter(analysis %in% "Y") %>%
      group_by(Year, Survey) %>% 
      summarise(N = n(),
                m = sum(Macaca_sur))  %>% 
      mutate(TypeName.1 = "Forest")
  )%>% 
  
  bind_rows(  #非森林(包括<50m & >=50m )的總計
    M.data %>% 
      filter(Year == 2024) %>% 
      filter(analysis %in% str_subset(analysis, "^Y")) %>%
      filter(analysis %in%
               str_subset(analysis, "month", negate = T)) %>%
      filter(TypeName.1 %in% "非森林") %>% 
      group_by(Year, Survey, TypeName.1) %>% 
      summarise(N = n(),
                m = sum(Macaca_sur))  
  )%>% 
  
  bind_rows(  #<50m 分別的小計
    M.data %>% 
      filter(Year == 2024) %>% 
      filter(analysis %in% str_subset(analysis, "^Y")) %>%
      filter(analysis %in%
               str_subset(analysis, "50m", negate = F)) %>%
 #     filter(Altitude < 50) %>%
      group_by(Year, Survey) %>% 
      summarise(N = n(),
                m = sum(Macaca_sur))  %>% 
      mutate(TypeName.1 = "less50")
  )%>% 
  
    bind_rows(  #total小計
    M.data %>% 
      filter(Year == 2024) %>% 
      filter(analysis %in% str_subset(analysis, "^Y")) %>%
      filter(analysis %in%
               str_subset(analysis, "month", negate = T)) %>%
      group_by(Year, Survey) %>% 
      summarise(N = n(),
                m = sum(Macaca_sur)) %>% 
      mutate(TypeName.1 = "Total") 
  )%>%
  
  mutate(E = m/N) %>%
  ungroup() %>%
  
  group_by(TypeName.1) %>%
  summarise(Mean_N = mean(N),
            Se_N = sd(N)/sqrt(n()),
            Mean_m = mean(m),
            Se_m = sd(m)/sqrt(n()),
            Mean_E = mean(E),
            Se_E = sd(E)/sqrt(n())) 

E_rate_Forest %>% 
  mutate(TypeName.1 = ordered(TypeName.1, c('Forest',
                                            '闊葉林',
                                       '竹林',
                                       '混淆林',
                                       '針葉林',
                                       'less50',
                                       '非森林',
                                       'Total'))) %>% 
  arrange(TypeName.1) %>% 
  mutate(Mean_E = round(Mean_E,3)) %>% 
  mutate(Se_E = round(Se_E,4)) %>% 
  flextable(.) %>%
  
  align(.,
        align = "center", part = "all") 



```

<br/>\
<br/>



```{r eval=FALSE, message=FALSE, warning=FALSE, include=FALSE}
# 2020~2024年各森林類型內的調查樣點數、臺灣獼猴( *Macaca cyclopis* )的猴群數及相對密度

E_rate_Forest<-

#森林內的分別計
M.data %>%  
  filter(analysis %in% "Y") %>%
  group_by(Year, Survey, TypeName.1) %>% 
  summarise(N = n(),
            m = sum(Macaca_sur)) %>% 
  
  bind_rows(  #森林的總計
    M.data %>% 
      filter(analysis %in% "Y") %>%
      group_by(Year, Survey) %>% 
      summarise(N = n(),
                m = sum(Macaca_sur))  %>% 
      mutate(TypeName.1 = "Forest")
  )%>% 
  
  bind_rows(  #非森林(包括<50m & >=50m )的總計
    M.data %>% 
      filter(analysis %in% str_subset(analysis, "^Y")) %>%
      filter(analysis %in%
               str_subset(analysis, "month", negate = T)) %>%
      filter(TypeName.1 %in% "非森林") %>% 
      group_by(Year, Survey, TypeName.1) %>% 
      summarise(N = n(),
                m = sum(Macaca_sur))  
  )%>% 
  
  bind_rows(  #<50m 分別的小計
    M.data %>% 
      filter(analysis %in% str_subset(analysis, "^Y")) %>%
      filter(analysis %in%
               str_subset(analysis, "50m", negate = F)) %>%
 #     filter(Altitude < 50) %>%
      group_by(Year, Survey) %>% 
      summarise(N = n(),
                m = sum(Macaca_sur))  %>% 
      mutate(TypeName.1 = "less50")
  )%>% 
  
    bind_rows(  #total小計
    M.data %>% 
      filter(analysis %in% str_subset(analysis, "^Y")) %>%
      filter(analysis %in%
               str_subset(analysis, "month", negate = T)) %>%
      group_by(Year, Survey) %>% 
      summarise(N = n(),
                m = sum(Macaca_sur)) %>% 
      mutate(TypeName.1 = "Total") 
  )%>%
  
  mutate(E = m/N) %>%
  ungroup() %>%
  
  group_by(TypeName.1) %>%
  summarise(Mean_N = mean(N),
            Se_N = sd(N)/sqrt(n()),
            Mean_m = mean(m),
            Se_m = sd(m)/sqrt(n()),
            Mean_E = mean(E),
            Se_E = sd(E)/sqrt(n())) 

E_rate_Forest %>% 
  mutate(TypeName.1 = ordered(TypeName.1, c('Forest',
                                            '闊葉林',
                                       '針葉林',
                                       '竹林',
                                       '混淆林',
                                       'less50',
                                       '非森林',
                                       'Total'))) %>% 
  arrange(TypeName.1) %>% 
  mutate_at(c('Mean_N', 'Se_N', 'Mean_m', 'Se_m'), function(x)round(x,1)) %>% 
  mutate(Mean_E = round(Mean_E,4)) %>% 
  mutate(Se_E = round(Se_E,4)) %>% 
  flextable(.) %>%
  
  align(.,
        align = "center", part = "all") 



```

<br/>\
<br/>


\newpage

Year的Encounter_rate

```{r echo=FALSE, message=FALSE, warning=FALSE}
Year.d.n <-
  M.data %>% 
#  filter(Year == 2023) %>% 
  filter(analysis %in% "Y") %>% 
  group_by(Survey, Year) %>% 
  summarise(N = n(), E = sum(Macaca_sur)/n()) %>% 
  
  group_by(Year) %>% 
  summarise(mean_N = mean(N) %>% round(1), 
            E25 = quantile(E,0.25),
            E50 = quantile(E,0.50),
            E75 = quantile(E,0.75)) %>% 
  ungroup()



M.data %>% 
#  filter(Year == 2023) %>% 
  filter(analysis %in% "Y") %>% 
  group_by(Survey, Year) %>% 
  summarise(E = sum(Macaca_sur)/n()) %>% 
  mutate(Year = as.character(Year)) %>% 
  
  ggplot(., aes(x = Year, y = E)) +
  geom_boxplot(size = 1, width = 0.3, fill= gray(.9))+
  geom_text(data = Year.d.n, 
            aes(x = 1:5,
                y = E75,
                label = mean_N),
            size = 4,
            hjust = -0.2, 
            vjust = -1,
            position = position_dodge(0.9), show.legend = F)+
  scale_y_continuous(limits = c(0,0.1))+
  labs(x = "年份", y = "相對密度 (群/樣點)")+
  theme(
    text = element_text(family="Microsoft JhengHei"),
    panel.border = element_rect(size = 1.5,fill = NA),
    panel.grid = element_blank(),
    panel.background = element_blank(),
    axis.text = element_text(size = 14,colour = "black"),
    axis.title = element_text(size = 18,colour = "black",
                              vjust = -2, hjust = 0.5),
    axis.title.y = element_text(angle = 270,
                                vjust = -1, hjust = 0.5),
    legend.position = "top"
  )
```

<br/>\
<br/>


\newpage

分署的Encounter_rate

```{r echo=FALSE, message=FALSE, warning=FALSE}
Office.d.n <-
  M.data %>% 
#  filter(Year == 2023) %>% 
  filter(analysis %in% "Y") %>% 
  group_by(Office,Survey, Year) %>% 
  summarise(N = n(), E = sum(Macaca_sur)/n()) %>% 
  
  group_by(Office) %>% 
  summarise(mean_N = mean(N) %>% round(1), 
            E25 = quantile(E,0.25),
            E50 = quantile(E,0.50),
            E75 = quantile(E,0.75)) %>% 
  ungroup()



M.data %>% 
#  filter(Year == 2023) %>% 
  filter(analysis %in% "Y") %>% 
  group_by(Office, Survey, Year) %>% 
  summarise(E = sum(Macaca_sur)/n()) %>% 
  mutate(Year = as.character(Year)) %>% 
  
  ggplot(., aes(x = Office, y = E)) +
  geom_boxplot(size = 1, width = 0.4, fill= gray(.9))+
  geom_text(data = Office.d.n, 
            aes(y = E75,
                label = mean_N),
            size = 3,
            hjust = -0.1, 
            vjust = -1,
            position = position_dodge(0.9), show.legend = F)+
  scale_y_continuous(breaks = seq(0,0.15,0.05), limits = c(0,0.15))+
  scale_x_discrete(limits = c( "新竹", "臺中", "南投","宜蘭",
                            "嘉義", "屏東", "花蓮", "臺東"))+
  labs(x = "分署", y = "相對密度 (群/樣點)")+
  theme(
    text = element_text(family="Microsoft JhengHei"),
    panel.border = element_rect(size = 1.5,fill = NA),
    panel.grid = element_blank(),
    panel.background = element_blank(),
    axis.text = element_text(size = 14,colour = "black"),
    axis.title = element_text(size = 18,colour = "black",
                              vjust = -2, hjust = 0.5),
    axis.title.y = element_text(angle = 270,
                                vjust = -1, hjust = 0.5),
    legend.position = "top"
  )
```

<br/>\
<br/>

\newpage

森林類型的Encounter_rate

```{r echo=FALSE, message=FALSE, warning=FALSE}
TypeName.d.n <-
  M.data %>% 
#  filter(Year == 2023) %>% 
  filter(analysis %in% "Y") %>% 
  group_by(TypeName.1,Survey, Year) %>% 
  summarise(N = n(), E = sum(Macaca_sur)/n()) %>% 
  
  group_by(TypeName.1) %>% 
  summarise(mean_N = mean(N) %>% round(1), 
            E25 = quantile(E,0.25),
            E50 = quantile(E,0.50),
            E75 = quantile(E,0.75)) %>% 
  ungroup()


M.data %>% 
#  filter(Year == 2023) %>% 
  filter(analysis %in% "Y") %>% 
  group_by(TypeName.1, Survey, Year) %>% 
  summarise(E = sum(Macaca_sur)/n()) %>% 
    mutate(Year = as.character(Year)) %>%
  
  ggplot(., aes(x = TypeName.1, y = E)) +
  geom_boxplot(size = 1, width = 0.4, fill= gray(.9))+
  geom_text(data = TypeName.d.n, 
            aes(y = E75,
                label = mean_N),
            size = 4,
            hjust = -0.2, 
            vjust = -1,
            position = position_dodge(0.9), show.legend = F)+

  scale_y_continuous(breaks = seq(0,0.1,0.02), 
                     limits = c(0,0.1))+
  scale_x_discrete(limits = c("針葉林", "混淆林", "竹林","闊葉林"))+
#  labs(x = "森林類型", y = "Encounter_rate")+
  labs(x = "森林類型", y = "相對密度(群/樣點)")+
  theme(
    text = element_text(family="Microsoft JhengHei"),
    panel.border = element_rect(size = 1.5,fill = NA),
    panel.grid = element_blank(),
    panel.background = element_blank(),
    axis.text = element_text(size = 14,colour = "black"),
    axis.title = element_text(size = 18,colour = "black",
                              vjust = -2, hjust = 0.5),
    axis.title.y = element_text(angle = 270,
                                vjust = -1, hjust = 0.5),
    legend.position = "top"
  )

```

<br/>\
<br/>

\newpage

Altitude的Encounter_rate

```{r message=FALSE, warning=FALSE, include=FALSE}
Alt.d.n <-
  M.data %>% 
#  filter(Year == 2023) %>% 
  filter(analysis %in% "Y") %>% 
  mutate(Altitude_f =cut(Altitude,
                         breaks = c(seq(0,4000,500)),
                         labels = c(seq(250,3750,500)),
                         include.lowest = T) ) %>% 
  group_by(Altitude_f,Survey, Year) %>% 
  summarise(N = n(), E = sum(Macaca_sur)/n()) %>% 
  
  group_by(Altitude_f) %>% 
  summarise(mean_N = mean(N) %>% round(1), 
            E25 = quantile(E,0.25),
            E50 = quantile(E,0.50),
            E75 = quantile(E,0.75)) %>% 
  ungroup()

  M.data %>%
 #   filter(Year == 2023) %>% 
  filter(analysis %in% "Y") %>% 
    mutate(Altitude = as.numeric(Altitude)) %>% 
  mutate(Altitude_f =cut(Altitude,
                         breaks = c(seq(0,4000,500)),
                         labels = c(seq(250,3750,500)),
                         include.lowest = T) ) %>% 
  group_by(Altitude_f,Survey, Year) %>% 
  summarise(N = n(), E = sum(Macaca_sur)/n()) %>% 
  flextable(.) %>%
  
  align(.,
        align = "center", part = "all")
```

```{r echo=FALSE, message=FALSE, warning=FALSE}
M.data %>% 
#   filter(Year == 2023) %>% 
 filter(analysis %in% "Y") %>%
  mutate(Altitude_f =cut(Altitude,
                         breaks = c(seq(0,4000,500)),
                         labels = c(seq(250,3750,500)),
                         include.lowest = T) ) %>% 
   
  group_by(Altitude_f,Survey, Year) %>% 
  summarise(E = sum(Macaca_sur)/n()) %>% 
  ungroup()%>% 
      mutate(Year = as.character(Year)) %>% 
  
  ggplot(., aes(x = Altitude_f, y = E)) +
  geom_boxplot(size = 1, width = 0.4, fill= gray(.9))+
  geom_text(data = Alt.d.n, 
            aes(y = E75,
                label = mean_N),
            size = 4,
            hjust = -0.2, 
            vjust = -1,
            position = position_dodge(0.9), show.legend = F)+
  
  scale_y_continuous(breaks = seq(0,1,0.2), 
                     limits = c(0,1)
                     )+ 

  labs(x = "Elevation (m)",
      y = "Encounter rate (troop/point)") +
  theme(
    text = element_text(family="Microsoft JhengHei"),
    panel.border = element_rect(size = 1.5,fill = NA),
    panel.grid = element_blank(),
    panel.background = element_blank(),
    axis.text = element_text(size = 14,colour = "black"),
    axis.title = element_text(size = 18,colour = "black",
                              vjust = -2, hjust = 0.5),
    legend.position = "top"
  )

```

<br/>

排除3000m以上，再畫一次

```{r echo=FALSE, message=FALSE, warning=FALSE}
M.data %>% 
#   filter(Year == 2023) %>% 
 filter(analysis %in% "Y") %>%
  mutate(Altitude_f =cut(Altitude,
                         breaks = c(seq(0,4000,500)),
                         labels = c(seq(250,3750,500)),
                         include.lowest = T) ) %>% 
   
  group_by(Altitude_f,Survey, Year) %>% 
  summarise(E = sum(Macaca_sur)/n()) %>% 
  ungroup()%>% 

  
  ggplot(., aes(x = Altitude_f, y = E)) +
  geom_boxplot(size = 1, width = 0.4, fill= gray(.9))+
  geom_text(data = Alt.d.n, 
            aes(y = E75,
                label = mean_N),
            size = 4,
            hjust = -0.2, 
            vjust = -1,
            position = position_dodge(0.9), show.legend = F)+
  
  scale_y_continuous(breaks = seq(0,0.10,0.02), 
                     limits = c(0,0.10)
                     )+ 
  coord_cartesian(xlim = c(1, 6))+
  labs(x = "Elevation (m)",
      y = "Encounter rate (troop/point)") +
  theme(
    text = element_text(family="Microsoft JhengHei"),
    panel.border = element_rect(size = 1.5,fill = NA),
    panel.grid = element_blank(),
    panel.background = element_blank(),
    axis.text = element_text(size = 14,colour = "black"),
    axis.title = element_text(size = 18,colour = "black",
                              vjust = -2, hjust = 0.5),
    legend.position = "top",
  )

```

<br/>\
\newpage   

```{r message=FALSE, warning=FALSE, include=FALSE}

  M.data %>% 
  filter(Year == 2024) %>% 
    filter(analysis %in% "Y") %>% 
    mutate(Altitude_f =cut(Altitude,
                           breaks = c(seq(0,4000,500)),
                           labels = c(seq(250,3750,500)),
                           include.lowest = T) ) %>% 
    group_by(Altitude_f,Survey, Year) %>% 
    summarise(E = sum(Macaca_sur)/n()) %>% 
    group_by(Altitude_f) %>% #確認中位數
    summarise(mid = median(E))

```

\newpage

#### 其他補充圖表

非森林的樣點離森林的距離 ( 包含海拔小於50m的樣點)

```{r echo=FALSE, message=FALSE, warning=FALSE}

M.data_notForest <- 
M.data %>%  
  filter(Year == 2024) %>% 
  filter(analysis %in% str_subset(analysis, "^Y")) %>%
  filter(analysis %in% str_subset(analysis, "month", negate = T)) %>%
  filter(TypeName.1 %in% "非森林") %>% 
  
  mutate(cut = cut(Distance, breaks = seq(20,240,10),
                   include.lowest = T, right = TRUE))

ggplot(M.data_notForest, aes(x = Distance))+
  geom_histogram(fill = gray(0.8), col = "black", breaks = seq(20,240,10))+
  stat_bin( geom="text", colour="black", size=3.5, breaks = seq(20,240,10),
           aes(label=..count.., y=(..count..)+0.6))+
  scale_x_continuous(breaks = seq(20,240,20),expand = c(0,0,0,5))+
  scale_y_continuous(expand = c(0,0,0,5))+
  labs(x = "Distance", y = "Count")+
  theme(
    text = element_text(family="serif"),
    panel.grid = element_blank(),
    panel.background = element_blank(),
    axis.text = element_text(size = 14,colour = "black"),
    axis.title = element_text(size = 18,colour = "black",
                              vjust = -2, hjust = 0.5)
  )
 

```

<br/>\
猴群分布的森林類型

```{r echo=FALSE, message=FALSE, warning=FALSE}
# 
Composition_Forest<- 
  M.data %>% 
  filter(Year == 2024) %>% 
   filter(analysis %in% "Y") %>% 
   group_by(TypeName.1,TypeName) %>% 
   summarise(N = n(), m = sum(Macaca_sur))

Composition_Forest %>% 
  flextable(.) %>% 
   width(., width = c(1,2,1,1))  %>%
  
  align(.,
        align = "center", part = "all") %>% 
  set_header_labels(
    'TypeName.1' = "森林類型",
    'TypeName' = "森林類型(4th森林圖層)",
    'N' = "樣點數",
    'm' = "猴群數" )

```

\newpage

猴群分布的海拔 (max Elevation = `r max(M.data[M.data$Year == 2023 & M.data $Macaca_sur==1 & M.data$analysis == "Y", ]$Altitude, na.rm = T)`)

```{r echo=FALSE, message=FALSE, warning=FALSE}
tmp <- 
M.data %>% 
  filter(Year == 2024) %>% 
    filter(analysis %in% "Y") %>%
  mutate(Macaca_sur = factor(Macaca_sur,
                         levels = c("1", "0")))%>% 
    mutate(Altitude_f =cut(Altitude,
                           breaks = c(seq(0,4000,200)),
                           labels = c(seq(100,3900,200)),
                           include.lowest = T) )  


ggplot(data = tmp, 
       aes(x = Altitude_f,
           color = Macaca_sur,
           fill = Macaca_sur))+
    geom_bar(position="stack")+
  labs(x = "Elevation")+
  theme(
    axis.text.x = element_text(angle = 270, vjust =0),
    legend.position = "top",
    axis.line = element_line(color = "black"),
    panel.background = element_blank()
  )

```

\newpage

(刪除重複猴群前)大於兩群的樣區列表

```{r echo=FALSE, message=FALSE, warning=FALSE}
####
Group_than2<- M.data %>%
  filter(Year == 2024) %>% 
   filter(analysis %in% "Y") %>% 
  filter(!Macaca_dist %in% "C") %>% 
   filter(Macaca_sur.ori %in% 2) %>% 
   group_by(Year, Survey, Site_N, Name) %>% 
     summarise(N_befor = n(),
               N_after = sum(Macaca_sur)) %>% 
   filter(N_befor>2)

Group_than2 %>% 
  left_join(M.data) %>% 
  select(Year,Station, Site_N, Name, Survey, Surveyor, N_befor, N_after) %>% 
  unique() %>% 
  mutate(Year = as.character(Year)) %>%
  mutate(Survey = as.character(Survey)) %>%
  
  reshape2::melt(id = 1:5) %>% 
  

  reshape2::dcast(Year+ Station + Site_N+Name ~ Survey + variable,
                  value.var = "value") %>% 
  arrange(Site_N) %>% 
  flextable(.) %>% 
#   width(., width = c(1,1,2,2.5,3,1,1,3,1,1)) %>% 
  
  add_header_row(., 
                 values = c(" "," "," "," ","第1旅次", "第2旅次"),
                 colwidths = c(1,1,1,1,3,3)) %>% 
  
  align(.,
        align = "center", part = "all")  %>% 


  set_header_labels(
    'Name' = "樣區名稱",
    'Site_N' = "樣區編號",
    '1_Surveyor' = "調查者",
    '1_N_befor' = "刪除前",
    '1_N_after' = "刪除後",
    '2_Surveyor' = "調查者",
    '2_N_befor' = "刪除前",
    '2_N_after' = "刪除後" ) 

```

單位：群

\newpage

```{r message=FALSE, warning=FALSE, include=FALSE}
 #可分析的樣點的海拔分布  
  M.data %>% 
  filter(Year == 2024) %>% 
    filter(analysis %in% "Y") %>% 
    mutate(Altitude_f =cut(Altitude,
                           breaks = c(seq(0,4000,100)),
                           labels = c(seq(50,4000,100)),
                           include.lowest =T) ) %>% 
    group_by(Altitude_f, Year) %>% 
    summarise(N = n()) %>% 

    ggplot(., aes(x = Altitude_f, y = N))+
    geom_point()
  
  
```

```{r eval=FALSE, message=FALSE, warning=FALSE, include=FALSE}
 
M.data %>% 
  filter(Year == 2024) %>% 
  filter(!is.na(Macaca_sur)) %>% 
    filter(!analysis %in%  str_subset(analysis, "^Y") ) %>% 
  select(Office,analysis, Site_N, Name, Surveyor, Station, Survey) %>% 
  unique() %>% 
  group_by(Office, Station, Surveyor, Site_N,Name) %>% 
  summarise(
    pp = paste0(analysis, collapse = ", ") %>% 
      str_split(., ", ", simplify = TRUE) %>%
      .[1,] %>% unique() %>% 
      paste0(collapse = ", ") ,
    xx = paste0(Survey, collapse = ", ") %>% 
      str_split(., ", ", simplify = TRUE) %>%
      .[1,] %>% unique() %>% 
      paste0(collapse = ", ") 
  ) %>% 
  mutate(pp = str_replace_all(pp, "6min", "未滿6分鐘")) %>%
  mutate(pp = str_replace_all(pp, "Toolate", "晚於11時")) %>%
  mutate(pp = str_replace_all(pp, "locate", "位置偏離")) %>%
  mutate(pp = str_replace_all(pp, "7day", "超過7日完成")) %>%
  mutate(pp = str_replace_all(pp, ",", "、")) %>%
  arrange(Office, Station, Surveyor) %>% 
  setNames(., c("分署","工作站", "調查者","樣區編號","樣區名稱","疏失類型", "旅次"))# %>% 
#  writexl::write_xlsx(., here("./Report of Foresty_20211124/調查疏失名單.xlsx"))



```

\newpage

2024年位在海拔50 m以上森林的調查樣點和臺灣獼猴(Macaca cyclopis)的猴群分布圖。紅色點( )為有猴群的樣點，藍色點( )為沒有猴群紀錄的樣點。

\newpage

2024年位在海拔50 m以上分署轄內的調查樣點和臺灣獼猴(Macaca cyclopis)的猴群分布圖。紅色點( )為有猴群的樣點，藍色點( )為沒有猴群紀錄的樣點。
